Performance testing pod struct with no binary writeout.
We are significantly faster than boost (up to about twice as fast) when not just dumping binary data.  Size seems to be
negligable between the two.  Speed goes up as data gets larger.  Boost's overhead for metadata is actually pretty small
when the data types themselves are not small.  For very small data types, it might start to make a difference.