minimal serialize functions now take Archive as const ref parameter
Even though there is currently no need to use this archive parameter, it is now
passed to the various minimal serialization functions instead of simply being
a template parameter.

The reason for this was ultimately allow for correct name-lookup of the functions
using ADL when the type was unable to provide this (e.g. a generic serialize function
for enums written in the cereal namespace).

Also fixed some issues with GCC 4.7 and the various traits that warn if the type for the
load isn't passed by const ref.  Had to use a workaround found here:

http://gcc.gnu.org/bugzilla/show_bug.cgi?id=51213 (=delete instead of private).