Implemented MinimalJSONOutputArchive
RapidJSON supports a non-whitespace version of the writer.
This commit changes the current JSONOutputArchive to exist inside
the details namespace (and be renamed as JSONOutputArchiveImpl)
and defines 2 usings for JSONOutputArchive and MinimalJSONOutputArchive.

JSONOutputArchive should support everything that the old one does,
Minimal has a slightly less featureful version of Options (since
indent and indent count aren't necessary) but prints without
unnecessary whitespace

Signed-off-by: Erich Keane <erich.keane@verizon.net>