Provide manual implementations for to_string() & friends
By defining CEREAL_IMPLEMENT_TO_STRING via the preprocessor, self-implemented
versions of the following STL functions will be provided.

* std::strtold()
* std::to_string()
* std::stoi()
* std::stol()
* std::stoll()
* std::stoul()
* std::stoull()
* std::stof()
* std::stod()
* std::stold()

The specific use case for this is using Android NDK with C++11 or higher
and on the GCC 4.9 + GNU STL toolchain/library. At the moment, implementations
of these libraries are not provided by this toolchain.