README: mention that library is not thread-safe
It's such a big deal, it should be among the first things a person gotta
know! I certainly didn't, it led to hours of frustration for why when I
have a bunch of clients communicating with a server (a thread per
client), sometimes I'm getting packets duplicated, even though clients
sent them correct. It costed lots of energy to pin down to
deserialization, and no, the CEREAL_THREAD_SAFE didn't help — probably
because, per the linked docs, you still not allowed to access the same
archive from multiple threads.

Signed-off-by: Konstantine Kharlamov <Hi-Angel@yandex.ru>