Scope cpp macros to avoid collisions w/client code
This public include file temporarily defines cpp macros `COPY`, `TRANS`, and `TAIL`.   Which means that if the client already has a definition of any of those, it will result in a compiler warning (or error) due to the duplicate definition, and then the client's macro will be undefined.   (Not just a hypothetical problem: I encountered a collision in practice, when using cereal with another library which also defined a acro `TRANS` without scoping.)

Simple fix of course: prefixing them all with `_CEREAL_`  (I used the leading underscore to self-document that this is an internal macro)